backend/app/core/config.py
from pydantic import BaseModel
import os

class Settings(BaseModel):
    db_host: str = os.getenv("DB_HOST", "localhost")
    db_port: int = int(os.getenv("DB_PORT", "3306"))
    db_name: str = os.getenv("DB_NAME", "portal_clientes")
    db_user: str = os.getenv("DB_USER", "root")
    db_pass: str = os.getenv("DB_PASS", "")
    jwt_secret: str = os.getenv("JWT_SECRET", "change-this")
    jwt_algorithm: str = "HS256"
    access_token_exp_minutes: int = 60 * 24
    upload_dir: str = os.getenv("UPLOAD_DIR", "/app/uploads")

settings = Settings()
